refactored code and layout files

+++ AndroidFitness/app/src/main/java/com/example/androidfitness/datalayer/macronutrients/MacronutrientsDB.java
@@ -4,7 +4,7 @@ import android.database.Cursor;
 
 public interface MacronutrientsDB {
 
-    public boolean addMacronutrients(int protein, int fat, int carbohydrates, int fibre, int salt);
+    public boolean addMacronutrients(float protein, float fat, float carbohydrates, float fibre, float salt);
 
     public Cursor macronutrientsReadAllData();
